requirements.txt
asgiref==3.4.1
Django==3.2.6
django-ckeditor==6.1.0
django-cleanup==5.2.0
django-js-asset==1.2.2
Pillow==8.3.1
pytz==2021.1
six==1.16.0
sorl-thumbnail==12.7.0
sqlparse==0.4.1
stats==0.1.2a0
